Finally! Drivers!
I have meant to do this for a while. VCore now uses a driver model to
manage data storage. By default, nothing really changed. Under the hood,
the data storage is now completely separate from the main bulk of VCore.

Please don't attempt to change drivers yet. VCore and many modules still
depend on direct access to the data tables and will break if the tables
are not correctly filled by the Data driver. As well as this, all
functions inside the tmysql4 driver are currently noop.